Circuit Construction and Usage

The output of typical Quantum simulation is the expectation values of the output quantum state based on certain measurement operators. In Qaintessent.jl, this is accomplished by applying a Quantum Circuit(s) consisting of a CircuitGateChain and Measurement Operators. See the corresponding sections for more information.

CircuitGate

The basic building block of any CircuitGateChain is CircuitGates. These are constructed from basic quantum Gates. The constructor for CircuitGates is complicated: CircuitGate Helper Functions and a CircuitGate Example can be found below.

Qaintessent.CircuitGate — Type

CircuitGate{M,N,G} <: AbstractCircuitGate{N}

Unitary quantum circuit gate. M is the number of wires affected by the CircuitGate, N is the overall number of quantum "wires" of the circuit, G is the basic gate used to construct the CircuitGate.

Qaintessent.CircuitGate — Method

CircuitGate{M,N,G}(iwire::NTuple{M,<:Integer}, gate::G) where {M,N,G}

Creates a CircuitGate{M,N,G} object. M is the number of wires affected by the CircuitGate, N is the overall number of quantum "wires" of the circuit, G is the basic gate used to construct the CircuitGate.

Qaintessent.CircuitGate — Method

CircuitGate(iwire::NTuple{M,<:Integer}, gate::AbstractGate{M}, N) where {M}

Creates a CircuitGate{M,N,G} object. M is the number of wires affected by the CircuitGate, N is the overall number of quantum "wires" of the circuit, G is the basic gate used to construct the CircuitGate.

Qaintessent.apply — Method

apply(cg::CircuitGate{M,N,G}, ψ::AbstractVector) where {M,N,G}

Apply a CircuitGate{M, N, G} to a quantum state vector ψ of N qubits.

CircuitGate Helper Functions

Qaintessent.single_qubit_circuit_gate — Method

single_qubit_circuit_gate(iwire::Integer, gate::AbstractGate{1}, N::Integer)

Construct a CircuitGate{1,N,G} object of basic gate type gate affecting wire iwire.

Qaintessent.two_qubit_circuit_gate — Method

two_qubit_circuit_gate(iwire1::Integer, iwire2::Integer, gate::AbstractGate{2}, N::Integer)

Construct a CircuitGate{2,N,G} object of basic gate type gate affecting wires iwire1 and iwire2.

Qaintessent.controlled_circuit_gate — Function

controlled_circuit_gate(itarget::Integer, icntrl::Union{Integer, Expr}, U::AbstractGate{1}, N::Integer)

Construct a CircuitGate{2,N,G} object of basic gate type U controlled by wire or Expr icntrl and affecting wire itarget.

controlled_circuit_gate(itarget::NTuple{M,<:Integer}, icntrl::Integer, U::AbstractGate{M}, N::Integer) where {M}

Construct a CircuitGate{M+1,N,G} object of basic gate type U controlled by wire or Expr icntrl and affecting wires in tuple itarget.

controlled_circuit_gate(itarget::Integer, icntrl::NTuple{K, Union{Int, Expr}}, U::AbstractGate{1}, N::Integer)  where {K}

Construct a CircuitGate{K+1,N,G} object of basic gate type U controlled by wires or Expr in tuple icntrl and affecting wire itarget.

controlled_circuit_gate(itarget::NTuple{M, <:Integer}, icntrl::NTuple{K, <:Union{Integer, Expr}}, U::AbstractGate{M}, N::Integer) where {K,M}

Construct a CircuitGate{M+K,N,G} object of basic gate type U controlled by wires in tuple icntrl and affecting wires in tuple itarget.

controlled_circuit_gate(itarget::NTuple{M, <:Integer}, reg::Register, U::AbstractGate{M}, N::Integer, ccntrl::AbstractVector{Integer}=Int[]) where {K,M}

Construct a CircuitGate{M+K,N,G} object of basic gate type U controlled by wires in register reg and affecting wires in tuple itarget.

controlled_circuit_gate(qreg::QRegister, icntrl::NTuple{M,<:Integer}, U::AbstractGate{M}, N::Integer) where {K,M}

Construct a CircuitGate{M+K,N,G} object of basic gate type U affecting wires in quantum register qreg and controlled by wires in tuple icntrl.

controlled_circuit_gate(qreg1::QRegister, qreg2::QRegister, U::AbstractGate{M}, N::Integer) where {K,M}

Construct a CircuitGate{M+K,N,G} object of basic gate type U controlled by wires in quantum register qreg and affecting wires in tuple itarget.

CircuitGate Example

The Qaintessent.matrix function can be used to convert CircuitGate objects to a CSC sparse matrix representation

    using Qaintessent
    N = 2
    cnot = controlled_circuit_gate(1, 2, XGate(), N)

    Qaintessent.matrix(cnot)

4×4 SparseArrays.SparseMatrixCSC{Float64,Int64} with 4 stored entries:
  [1, 1]  =  1.0
  [2, 2]  =  1.0
  [4, 3]  =  1.0
  [3, 4]  =  1.0

The CircuitGate can then be applied to a quantum state in state vector form.

    ψ = [0, 0, 1, 0]

    apply(cnot, ψ)

4-element Array{Int64,1}:
 0
 0
 0
 1

Moments

When performing a quantum simulation, it may be required to define an intermediate state; this can be used to optimize the mapping of gates to physical qubits. In Qaintessent.jl, we allow for groupings of CircuitGate(s) that can be run in parallel into Moments. A Moment Example is shown below.

Qaintessent.Moment — Type

Moment{N}

Represents an intermediate state within a given circuit. N is the overall number of quantum "wires" of the circuit.

Qaintessent.Moment — Method

Moment{N}(g::AbstractCircuitGate{N}) where {N}

Create a Moment{N} object consisting of a single CircuitGate{N} object.

Moment{N}(g::AbstractVector{<:AbstractCircuitGate{N}}) where {N}

Create a Moment{N} object consisting of multiple CircuitGate{N} objects.

Missing docstring.

Missing docstring for apply(m::Moment{N}, ψ::AbstractVector) where {N}. Check Documenter's build log for details.

Moment Example

    using Qaintessent

    N = 3
    m = Moment{N}(
        [
            controlled_circuit_gate(1, 2, XGate(), N),
            single_qubit_circuit_gate(3, RxGate(0.2π), N),
        ]
    )

CircuitGate{2,3,ControlledGate{1,2}}((1, 2), ControlledGate{1,2}(XGate()), Union{Int64, Expr}[])
CircuitGate{1,3,RxGate}((3,), RxGate([0.6283185307179586]), Int64[])

The Moment can then be applied to a quantum state in state vector form.

    ψ = randn(ComplexF64, 2^N)

    apply(m, ψ)

8-element Array{Complex{Float64},1}:
 -0.02414209929187225 - 0.04110467550090291im
 -0.29574281977766786 + 0.6730647462830401im
  -1.2249153395610646 + 0.2687442525171615im
   -0.519954350759259 + 0.17825798553142455im
  -0.8600930488090999 - 0.43540082774331806im
  -0.9639675844951082 - 1.4407312285851328im
   0.2478327799581206 + 0.49134801240775866im
  0.05524338492558366 - 1.0759847650757697im

CircuitGateChain

CircuitGateChains can be constructed in two ways, either by providing a Vector of CircuitGate objects or Moment objects. A CircuitGateChain Example can be seen below.

Qaintessent.CircuitGateChain — Type

CircuitGateChain{N}

Chain of quantum circuit gates in a circuit of N qubits.

Qaintessent.CircuitGateChain — Method

CircuitGateChain{N}(gates::AbstractVector{<:AbstractCircuitGate{N}}) where {N}

Chain of quantum circuit gates in a circuit of N qubits. Constructed from vector of CircuitGate{N} objects.

CircuitGateChain{N}(moments::AbstractVector{<:AbstractMoment{N}}, creg::AbstractVector{BitArray{1}}=BitArray{1}[]) where {N}

Chain of quantum circuit gates in a circuit of N qubits, constructing from vector of Moment{N} objects.

CircuitGateChain{N}(cregs::AbstractVector{CRegister})

Chain of quantum circuit gates in a circuit of N qubits, constructing with vector of classical registers.

CircuitGateChain{N}()

Chain of quantum circuit gates in a circuit of N qubits.

Missing docstring.

Missing docstring for apply(c::CircuitGateChain{N}, ψ::AbstractVector) where {N}. Check Documenter's build log for details.

CircuitGateChain Example

In the following example, a 3-qubit Quantum Fourier Transform is applied to an arbitrary input state

    using Qaintessent

    N = 3
    cgc = CircuitGateChain{N}(
        [
            single_qubit_circuit_gate(1, HadamardGate(), N),
            controlled_circuit_gate(2, 1, PhaseShiftGate(2π/2^2), N),
            controlled_circuit_gate(3, 1, PhaseShiftGate(2π/2^3), N),
            single_qubit_circuit_gate(2, HadamardGate(), N),
            controlled_circuit_gate(3, 1, PhaseShiftGate(2π/2^2), N),
            single_qubit_circuit_gate(2, HadamardGate(), N),
        ]
    )

    3 —————————————[Pϕ]————————[Pϕ]———————
                    |           |
    2 ———————[Pϕ]————————[H ]————————[H ]—
              |     |           |
    1 —[H ]———•—————•———————————•—————————

The CircuitGateChain can then be applied to a quantum state in state vector form.

    ψ = randn(ComplexF64, 2^N)

    apply(cgc, ψ)

8-element Array{Complex{Float64},1}:
  1.0213633459314218 - 0.9247841882753862im
 -0.2995808928998076 + 0.403884313432495im
  0.9290335361782149 - 0.7628441463993755im
 0.03673583356158832 - 0.05492748556004257im
 0.04025182228087879 - 0.2633870610825533im
  0.9498588771532094 - 0.030332494427789437im
  0.4138080624099504 + 0.6898401519170307im
   0.469503739873149 - 0.9194016627894229im

Measurement Operators

Measurement operators can be defined in a MeasurementOps object. These are used in conjunction with a CircuitGateChain to create a Circuit object. Usage of MeasurementOps can be seen in the Circuit Example shown below.

Qaintessent.MeasurementOps — Type

MeasurementOps{N}

Pairwise commuting measurement operators (Hermitian matrices) for circuit of size N.

Qaintessent.MeasurementOps — Method

MeasurementOps{N}(mop::AbstractMatrix) where {N}

constructs a MeasurementOps{N} object for circuit of size N from single $2^{N} \times 2^{N}$ matrix.

MeasurementOps{N}(mops::AbstractVector{<:AbstractMatrix}) where {N}

constructs a MeasurementOps{N} object for a circuit of size N from a vector of $2^{N} \times 2^{N}$ matrices.

MeasurementOps{N}(cg::CircuitGate) where {N}

constructs a MeasurementOps{N} object for a circuit of size N from a single CircuitGate{N} object.

MeasurementOps{N}(cgs::AbstractVector{<:CircuitGate}) where {N}

constructs a MeasurementOps{N} object for a circuit of size N from a vector of CircuitGate{N} objects.

Circuit

Circuit objects combine a CircuitGateChain and a MeasurementOps object. Applying a Circuit to a given input quantum state outputs the various expectation values from the measurement operators defined in the MeasurementOps object. A simple circuit is shown in the Circuit Example.

Qaintessent.Circuit — Type

Circuit{N}

Quantum circuit consisting of a unitary gate chain and measurement operators.

Missing docstring.

Missing docstring for apply(c::Circuit{N}, ψ::AbstractVector) where {N}. Check Documenter's build log for details.

Circuit Example

    using Qaintessent
    N = 3
    cgc = CircuitGateChain{N}([
        single_qubit_circuit_gate(1, HadamardGate(), N),
        single_qubit_circuit_gate(2, HadamardGate(), N),
        single_qubit_circuit_gate(3, HadamardGate(), N),
    ])
    I = [1 0 ; 0 1]
    x = Qaintessent.matrix(XGate())
    meas_op = kron(kron(I,x), I)
    meas = MeasurementOps{N}([meas_op])
    c = Circuit{N}(cgc, meas)

    3 —[H ]—

    2 —[H ]—

    1 —[H ]—

Applying the Circuit object to a 3-qubit quantum state all in the ground 0 state.

    ψ = Float64[1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
    
    apply(c, ψ)