Density Matrix representation

Qaintessent.jl supports the manipulation of mixed states as DensityMatrix(s).

Density Matrices

Qaintessent.DensityMatrix — Type

DensityMatrix

Density matrix, represented with respect to identity and Pauli basis (σ_j/2) by storing 4 real coefficients for each qubit, i.e., a real vector of length 4^N for N qubits.

Missing docstring.

Missing docstring for apply(cgs::Vector{<:CircuitGate}, ρ::DensityMatrix). Check Documenter's build log for details.

Missing docstring.

Missing docstring for apply(c::Circuit{N}, ρ::DensityMatrix) where {N}. Check Documenter's build log for details.