Gradients

Qaintessent.jl can perform a backward pass through a circuit to compute gradients of a (fictitious) cost function with respect to the circuit parameters and input wavefunction. Since quantum gates are unitary, the intermediate quantum states can be re-computed on the fly during backwards traversal, i.e., need not be stored during the forward pass.

See also this article regarding complex differentiation.

Qaintessent.gradients — Function

gradients(c::Circuit{N}, ψ::AbstractVector, Δ::AbstractVector{<:Real}) where {N}

Perform a backward pass to compute gradients of a (fictitious) cost function with respect to the circuit parameters of c and input wavefunction ψ. Δ contains the cost function gradients with respect to the circuit outputs (measurement operator averages). The gradients with respect to the circuit parameters are returned in a duplicate circuit; the overall return value is the tuple (dc::Circuit{N}, dψ::AbstractVector).