Circuit Construction and Usage

The output of typical Quantum simulation is the expectation values of the output quantum state based on certain measurement operators. In Qaintessent.jl, this is accomplished by applying a Quantum Circuit(s) consisting of a series of CircuitGate objects and Measurement Operators. See the corresponding sections for more information.

CircuitGate

The basic building block of any Circuit is CircuitGates. These are constructed from basic quantum Gates. The constructor for CircuitGates is complicated: CircuitGate Helper Functions and a CircuitGate Example can be found below.

Qaintessent.CircuitGate — Type

CircuitGate{M,G} <: AbstractCircuitGate

Unitary quantum circuit gate. M is the number of wires affected by the CircuitGate, N is the overall number of quantum "wires" of the circuit, G is the basic gate used to construct the CircuitGate.

Qaintessent.apply — Method

apply(cg::CircuitGate{M,G}, ψ::Vector{<:Complex}) where {M,G}

Apply a CircuitGate{M,G} to a quantum state vector ψ.

CircuitGate Helper Functions

Qaintessent.circuit_gate — Function

circuit_gate

Construct a CircuitGate object from basic gate types.

CircuitGate Example

The Qaintessent.matrix function can be used to convert CircuitGate objects to a CSC sparse matrix representation

    using Qaintessent
    N = 2
    cnot = circuit_gate(1, XGate(), 2)

    Qaintessent.matrix(cnot)

4×4 SparseArrays.SparseMatrixCSC{Complex{Float64},Int64} with 4 stored entries:
  [1, 1]  =  1.0+0.0im
  [2, 2]  =  1.0+0.0im
  [4, 3]  =  1.0+0.0im
  [3, 4]  =  1.0+0.0im

The CircuitGate can then be applied to a quantum state in state vector form.

    ψ = ComplexF64[0, 0, 1, 0]

    apply(cnot, ψ)

4-element Array{Complex{Float64},1}:
 0.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im
 1.0 + 0.0im

Moments

When performing a quantum simulation, it may be required to define an intermediate state; this can be used to optimize the mapping of gates to physical qubits. In Qaintessent.jl, we allow for groupings of CircuitGate(s) that can be run in parallel into Moments. A Moment Example is shown below.

Qaintessent.Moment — Type

Moment

Represents an intermediate state within a given circuit.

Qaintessent.apply — Method

apply(m::Moment, ψ::Vector{<:Complex})

returns state vector of N qubits after applying a Moment{N} object to a quantum state vector of N qubits ψ

Moment Example

    using Qaintessent
    N = 3
    m = Moment(
        [
            circuit_gate(1, X, 2),
            circuit_gate(3, RxGate(0.2π)),
        ]
    )

CircuitGate{2,ControlledGate{XGate}}((1, 2), ControlledGate{XGate}(XGate(), 1))
CircuitGate{1,RxGate}((3,), RxGate([0.6283185307179586]))

The Moment can then be applied to a quantum state in state vector form.

    ψ = randn(ComplexF64, 2^N)

    apply(m, ψ)

8-element Array{Complex{Float64},1}:
   0.4751702297108598 + 0.2595980961408783im
   0.4092306016230728 - 1.0021665097491084im
 -0.12304089997557939 + 1.8766485003650792im
  0.31052299593527044 + 0.5420022733456987im
  0.46409755237251793 + 0.28311117013299664im
  -0.7180604514467264 - 0.2709397383396289im
   1.0938977925652855 - 0.5219507308225074im
 0.022727994456767514 + 0.750154247618473im

Measurement Operators

Measurement operators can be defined in a MeasurementOperator object. These are used in conjunction with a series ofCircuitGate objects to create a Circuit object. Usage of MeasurementOperator can be seen in the Circuit Example shown below.

Qaintessent.MeasurementOperator — Type

MeasurementOperator{M,G} <: AbstractCircuitGate

General complex operator. M is the number of wires affected by the CircuitGate, G is the basic gate used to construct the Operator.

Circuit

Circuit objects combine a Vector of Moment and a MeasurementOperator objects. Applying a Circuit to a given input quantum state outputs the various expectation values from the measurement operators defined in the MeasurementOperator objects. A simple circuit is shown in the Circuit Example.

Qaintessent.Circuit — Type

Circuit

Quantum circuit consisting of a unitary gate chain and measurement operators.

Qaintessent.apply — Method

apply(c::Circuit{N}, ψ::Vector{<:Complex}) where {N}

returns list of expectation values from measurement operators in c.meas after applying circuit gates in c.cgc on state vector of N qubits ψ

Circuit Example

    using Qaintessent
    N = 3
    cgc = CircuitGate[
        circuit_gate(1, HadamardGate()),
        circuit_gate(2, HadamardGate()),
        circuit_gate(3, HadamardGate()),
    ]
    I = ComplexF64[1 0 ; 0 1]
    meas = [MeasurementOperator(X, (1,))]
    c = Circuit{N}(cgc, meas)

    3 —[H ]—

    2 —[H ]—

    1 —[H ]—

Applying the Circuit object to a 3-qubit quantum state all in the ground 0 state.

    ψ = ComplexF64[1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]

    apply(c, ψ)

1-element Array{Float64,1}:
 0.9999999999999996