Circuit Construction and Usage

The output of typical Quantum simulation is the expectation values of the output quantum state based on certain measurement operators. In Qaintessent.jl, this is accomplished by applying a Quantum Circuit(s) consisting of a CircuitGateChain and Measurement Operators. See the corresponding sections for more information.

CircuitGate

The basic building block of any CircuitGateChain is CircuitGates. These are constructed from basic quantum Gates. The constructor for CircuitGates is complicated: CircuitGate Helper Functions and a CircuitGate Example can be found below.

Qaintessent.CircuitGate — Type

CircuitGate{M,N,G} <: AbstractCircuitGate{N}

Abstract unitary quantum circuit gate. M is the number of wires affected by the CircuitGate, N is the overall number of quantum "wires" of the circuit, G is the basic gate used to construct the CircuitGate.

Qaintessent.CircuitGate — Method

CircuitGate{M,N,G}(iwire::NTuple{M, <:Integer}, gate::G) where {M,N,G}

creates a CircuitGate{M,N,G} object. M is the number of wires affected by the CircuitGate, N is the overall number of quantum "wires" of the circuit, G is the basic gate used to construct the CircuitGate.

Qaintessent.CircuitGate — Method

CircuitGate(iwire::NTuple{M, <:Integer}, gate::AbstractGate{M}, N) where {M}

creates a CircuitGate{M,N,G} object. M is the number of wires affected by the CircuitGate, N is the overall number of quantum "wires" of the circuit, G is the basic gate used to construct the CircuitGate.

Qaintessent.apply — Method

apply(cg::CircuitGate{M,N,G}, ψ::AbstractVector) where {M,N,G}

returns state vector of N qubits after applying a CircuitGate{M, N, G} object to a quantum state vector of N qubits ψ.

CircuitGate Helper Functions

Qaintessent.single_qubit_circuit_gate — Method

single_qubit_circuit_gate(iwire::Integer, gate::AbstractGate{1}, N::Integer)

returns a CircuitGate{1,N,G} object of basic gate type gate affecting wire iwire.

Qaintessent.two_qubit_circuit_gate — Method

two_qubit_circuit_gate(iwire1::Integer, iwire2::Integer, gate::AbstractGate{2}, N::Integer)

returns a CircuitGate{2,N,G} object of basic gate type gate affecting wires iwire1 and iwire2.

Qaintessent.controlled_circuit_gate — Function

controlled_circuit_gate(icntrl::Integer, itarget::Integer, U::AbstractGate{1}, N::Integer)

returns a CircuitGate{2,N,G} object of basic gate type U controlled by wire icntrl and affecting wire itarget.

controlled_circuit_gate(icntrl::Integer, itarget::NTuple{M, <:Integer}, U::AbstractGate{M}, N::Integer) where {M}

returns a CircuitGate{M+1,N,G} object of basic gate type U controlled by wire icntrl and affecting wires in tuple itarget.

controlled_circuit_gate(icntrl::NTuple{K, <:Integer}, itarget::Integer, U::AbstractGate{1}, N::Integer)  where {K}

returns a CircuitGate{K+1,N,G} object of basic gate type U controlled by wires in tuple icntrl and affecting wire itarget.

controlled_circuit_gate(icntrl::NTuple{K, <:Integer}, itarget::NTuple{M, <:Integer}, U::AbstractGate{M}, N::Integer) where {K,M}

returns a CircuitGate{M+K,N,G} object of basic gate type U controlled by wires in tuple icntrl and affecting wires in tuple itarget.

CircuitGate Example

The Qaintessent.matrix function can be used to convert CircuitGate objects to a CSC sparse matrix representation

    using Qaintessent
    N = 2
    cnot = controlled_circuit_gate(1, 2, XGate(), N)

    Qaintessent.matrix(cnot)

4×4 SparseArrays.SparseMatrixCSC{Float64,Int64} with 4 stored entries:
  [1, 1]  =  1.0
  [2, 2]  =  1.0
  [4, 3]  =  1.0
  [3, 4]  =  1.0

The CircuitGate can then be applied to a quantum state in state vector form.

    ψ = [0, 0, 1, 0]

    apply(cnot, ψ)

4-element Array{Int64,1}:
 0
 0
 0
 1

Moments

When performing a quantum simulation, it may be required to define an intermediate state; this can be used to optimize the mapping of gates to physical qubits. In Qaintessent.jl, we allow for groupings of CircuitGate(s) that can be run in parallel into Moments. A Moment Example is shown below.

Qaintessent.Moment — Type

AbstractMoment{N}

Represents an intermediate state within a given circuit. N is the overall number of quantum "wires" of the circuit.

Qaintessent.Moment — Method

Moment{N}(g::AbstractCircuitGate{N}) where {N}

creates a Moment{N} object consisting of a single CircuitGate{N} object.

Moment{N}(g::AbstractVector{<:AbstractCircuitGate{N}}) where {N}

creates a Moment{N} object consisting of multiple CircuitGate{N} objects.

Qaintessent.apply — Method

apply(m::Moment{N}, ψ::AbstractVector) where {N}

returns state vector of N qubits after applying a Moment{N} object to a quantum state vector of N qubits ψ

Moment Example

    using Qaintessent

    N = 3
    m = Moment{N}(
        [
            controlled_circuit_gate(1, 2, XGate(), N),
            single_qubit_circuit_gate(3, RxGate(0.2π), N),
        ]
    )

CircuitGate{2,3,ControlledGate{1,2}}((1, 2), ControlledGate{1,2}(XGate()))
CircuitGate{1,3,RxGate}((3,), RxGate([0.6283185307179586]))

The Moment can then be applied to a quantum state in state vector form.

    ψ = randn(ComplexF64, 2^N)

    apply(m, ψ)

8-element Array{Complex{Float64},1}:
   0.5054948460289349 - 1.1386185836694582im 
  -0.4324365844538468 + 0.37861068108891627im
  0.12219563923226703 - 0.6829334836895642im 
    1.343212443068089 - 0.0872368231434957im 
  0.15155368009254483 + 0.7902246223900757im 
   0.9033987244119622 + 0.21142167360549158im
 -0.32329529037226323 + 0.684746082691335im  
   1.1973922654916955 - 0.24719429593164505im

CircuitGateChain

CircuitGateChains can be constructed in two ways, either by providing a Vector of CircuitGate objects or Moment objects. A CircuitGateChain Example can be seen below.

Qaintessent.CircuitGateChain — Type

CircuitGateChain{N}

Chain of quantum circuit gates in a circuit of N qubits.

Qaintessent.CircuitGateChain — Method

CircuitGateChain{N}(gates::AbstractVector{<:AbstractCircuitGate{N}}) where {N}

Chain of quantum circuit gates in a circuit of N qubits. Constructed from vector of CircuitGate{N} objects.

CircuitGateChain{N}(moments::AbstractVector{<:AbstractMoment{N}}) where {N}

Chain of quantum circuit gates in a circuit of N qubits, constructing from vector of Moment{N} objects.

Qaintessent.apply — Method

apply(cgc::CircuitGateChain{N}, ψ::AbstractVector) where {N}

returns state vector of N qubits after applying a CircuitGateChain{N} object to quantum state vector of N qubits ψ

CircuitGateChain Example

In the following example, a 3-qubit Quantum Fourier Transform is applied to an arbitrary input state

    using Qaintessent

    N = 3
    cgc = CircuitGateChain{N}(
        [
            single_qubit_circuit_gate(1, HadamardGate(), N),
            controlled_circuit_gate(2, 1, PhaseShiftGate(2π/2^2), N),
            controlled_circuit_gate(3, 1, PhaseShiftGate(2π/2^3), N),
            single_qubit_circuit_gate(2, HadamardGate(), N),
            controlled_circuit_gate(3, 1, PhaseShiftGate(2π/2^2), N),
            single_qubit_circuit_gate(2, HadamardGate(), N),
        ]
    )

    1 —[H ]——[Pϕ]——[Pϕ]————————[Pϕ]———————
              |     |           |
    2 ————————•——————————[H ]————————[H ]—
                    |           |
    3 ——————————————•———————————•—————————

The CircuitGateChain can then be applied to a quantum state in state vector form.

    ψ = randn(ComplexF64, 2^N)

    apply(cgc, ψ)

8-element Array{Complex{Float64},1}:
  -0.2919937837828615 + 0.3286790366066088im 
 -0.32553875228077755 - 0.5660076063143622im 
   0.5226378213090496 + 0.5050523430640134im 
   0.5575410674904546 - 0.2255492483938409im 
   -1.044339575189784 + 0.1282497690442185im 
  -1.1396342363741359 - 0.25438775159376054im
   0.3536351114704629 - 0.1024809730599007im 
   -1.092859803753081 + 0.16999007098136126im

Measurement Operators

Measurement operators can be defined in a MeasurementOps object. These are used in conjunction with a CircuitGateChain to create a Circuit object. Usage of MeasurementOps can be seen in the Circuit Example shown below.

Qaintessent.MeasurementOps — Type

MeasurementOps{N}

Pairwise commuting measurement operators (Hermitian matrices) for circuit of size N.

Qaintessent.MeasurementOps — Method

MeasurementOps{N}(mop::AbstractMatrix) where {N}

constructs a MeasurementOps{N} object for circuit of size N from single $2^{N} \times 2^{N}$ matrix.

MeasurementOps{N}(mops::AbstractVector{<:AbstractMatrix}) where {N}

constructs a MeasurementOps{N} object for a circuit of size N from a vector of $2^{N} \times 2^{N}$ matrices.

MeasurementOps{N}(cg::CircuitGate) where {N}

constructs a MeasurementOps{N} object for a circuit of size N from a single CircuitGate{N} object.

MeasurementOps{N}(cgs::AbstractVector{<:CircuitGate}) where {N}

constructs a MeasurementOps{N} object for a circuit of size N from a vector of CircuitGate{N} objects.

Circuit

Circuit objects combine a CircuitGateChain and a MeasurementOps object. Applying a Circuit to a given input quantum state outputs the various expectation values from the measurement operators defined in the MeasurementOps object. A simple circuit is shown in the Circuit Example.

Qaintessent.Circuit — Type

Circuit{N}

Quantum circuit consisting of a unitary gate chain and measurement operators.

Qaintessent.apply — Method

apply(c::Circuit{N}, ψ::AbstractVector) where {N}

returns list of expectation values from measurement operators in c.meas after applying circuit gates in c.cgc on state vector of N qubits ψ

Circuit Example

    using Qaintessent
    N = 3
    cgc = CircuitGateChain{N}([
        single_qubit_circuit_gate(1, HadamardGate(), N),
        single_qubit_circuit_gate(2, HadamardGate(), N),
        single_qubit_circuit_gate(3, HadamardGate(), N),
    ])
    I = [1 0 ; 0 1]
    x = Qaintessent.matrix(XGate())
    meas_op = kron(kron(I,x), I)
    meas = MeasurementOps{N}([meas_op])
    c = Circuit{N}(cgc, meas)

    1 —[H ]—

    2 —[H ]—

    3 —[H ]—

Applying the Circuit object to a 3-qubit quantum state all in the ground 0 state.

    ψ = Float64[1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]

    apply(c, ψ)

1-element Array{Float64,1}:
 0.9999999999999993