Circuit Construction and Usage

The output of typical Quantum simulation is the expectation values of the output quantum state based on certain measurement operators. In Qaintessent.jl, this is accomplished by applying a Quantum Circuit consisting of a series of CircuitGate objects and Measurement Operators.

Moments

When performing a quantum simulation, it may be required to define an intermediate state; this can be used to optimize the mapping of gates to physical qubits. In Qaintessent.jl, we allow for groupings of CircuitGate(s) that can be run in parallel into Moments. A Moment Example is shown below.

Qaintessent.Moment — Type

Moment

Represents an intermediate state within a given circuit.

Qaintessent.apply — Method

apply(m::Moment, ψ::Vector{<:Complex})

returns state vector of N qubits after applying a Moment{N} object to a quantum state vector of N qubits ψ

Moment Example

    using Qaintessent
    N = 3
    m = Moment(
        [
            circuit_gate(1, X, 2),
            circuit_gate(3, RxGate(0.2π)),
        ]
    )

CircuitGate{2,ControlledGate{XGate}}((1, 2), ControlledGate{XGate}(XGate(), 1))
CircuitGate{1,RxGate}((3,), RxGate([0.6283185307179586]))

The Moment can then be applied to a quantum state in state vector form.

    ψ = randn(ComplexF64, 2^N)

    apply(m, ψ)

8-element Array{Complex{Float64},1}:
   -0.3900786173826597 + 1.1347869596522346im
   -0.9597694895589257 - 1.3825157065173226im
    1.5728952665071758 + 0.13420386134085682im
    -0.590212421349418 + 0.35648735710802204im
      0.20116001945004 - 0.5305429594823867im
    0.7123367978192574 - 1.1550433753795568im
   -0.3703815866120241 + 0.9158294666288239im
 -0.024250549244820285 + 0.09960826537520868im

Measurement Operators

Measurement operators can be defined in a MeasurementOperator object. These are used in conjunction with a series ofCircuitGate objects to create a Circuit object. Usage of MeasurementOperator can be seen in the Circuit Example shown below.

Qaintessent.MeasurementOperator — Type

MeasurementOperator{M,G}

General measurement operator. M is the number of wires affected by the measurement operator, G is the actual operator type (gate or sparse matrix).

julia> m = MeasurementOperator(X, (1,))
MeasurementOperator{1,XGate}(XGate(), (1,))

julia> m = MeasurementOperator([0 1; 1 0], (1,))
MeasurementOperator{1,SparseArrays.SparseMatrixCSC{Complex{Float64},Int64}}(
  [2, 1]  =  1.0+0.0im
  [1, 2]  =  1.0+0.0im, (1,))

Circuit

Circuit objects combine a Vector{Moment} object and Vector{MeasurementOperator} object. Applying a Circuit to a given input quantum state outputs the various expectation values from the measurement operators defined in the Qaintessent.MeasurementOperator objects. A simple circuit is shown in the Circuit Example.

Qaintessent.Circuit — Type

Circuit{N}

Quantum circuit consisting of a unitary gate chain and measurement operators with N qubits.

Qaintessent.apply — Method

apply(c::Circuit{N}, ψ::Vector{<:Complex}) where {N}

returns list of expectation values from measurement operators in c.meas after applying circuit gates in c.cgc on state vector of N qubits ψ

Circuit Example

julia> N = 3;

julia> cgs = CircuitGate[circuit_gate(1, HadamardGate()), circuit_gate(2, HadamardGate()), circuit_gate(3, HadamardGate())];

julia> meas = [MeasurementOperator(X, (1,))]; # Measure with regard to X basis on first qubit.

julia> c = Circuit{N}(cgs, meas) # create circuit object

    3 —[H ]—
            
    2 —[H ]—
            
    1 —[H ]—

Applying the Circuit object to a 3-qubit quantum state all in the ground 0 state.

julia> ψ = ComplexF64[1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0];

julia> apply(c, ψ)
1-element Array{Float64,1}:
 0.9999999999999996