Circuit Gates

A CircuitGate object consists of an Qaintessent.AbstractGate and wire ordering information. This allows an arbitrary quantum operation can be encoded as vector of CircuitGate objects and applied to a given quantum state.

CircuitGate

Qaintessent.CircuitGate — Type

CircuitGate{M,G} <: AbstractCircuitGate

Unitary quantum circuit gate. M is the number of wires affected by the CircuitGate, and G is the basic gate used to construct the CircuitGate.

Qaintessent.circuit_gate — Function

circuit_gate

helper function to construct a CircuitGate object from basic gate types.

julia> circuit_gate(1, X)
CircuitGate{1,XGate}((1,), XGate())

julia> circuit_gate(1, X, 2)
CircuitGate{2,ControlledGate{XGate}}((1, 2), ControlledGate{XGate}(XGate(), 1))

julia> circuit_gate((1,2), SwapGate(), (3,4))
CircuitGate{4,ControlledGate{SwapGate}}((1, 2, 3, 4), ControlledGate{SwapGate}(SwapGate(), 2))

Qaintessent.apply — Method

apply(ψ::Vector{<:Complex}, cg::CircuitGate{M,G}) where {M,G}

Apply a CircuitGate to a quantum state vector ψ.

Examples

julia> cg = circuit_gate(1, HadamardGate());
julia> ψ = [1 0];
julia> apply(ψ, cg)
2-element Array{Complex{FloatQ},1}:
 0.7071067811865475 + 0.0im
 0.7071067811865475 + 0.0im

Qaintessent.apply — Method

apply(ψ::Vector{<:Complex}, cgs::Vector{<:CircuitGate})

Apply a sequence of CircuitGate(s) to a quantum state vector ψ.

Examples

julia> cgs = [circuit_gate(1, HadamardGate()),
                circuit_gate(1, X),
                circuit_gate(1, Y)];
julia> ψ = [1 0];
julia> apply(ψ, cgs)
2-element Array{Complex{FloatQ},1}:
 0.0 - 0.7071067811865475im
 0.0 + 0.7071067811865475im

Qaintessent.sparse_matrix — Method

sparse_matrix(cg::CircuitGate{M,G}, N::Integer=0) where {M,G <: AbstractGate}

returns matrix representation of a CircuitGate object that can applied to a state vector of N qubits. N can be

Examples

julia> cg = circuit_gate(2, Y, 1);
julia> sparse_matrix(cg)
4×4 SparseArrays.SparseMatrixCSC{Complex{FloatQ},Int64} with 4 stored entries:
  [1, 1]  =  1.0+0.0im
  [4, 2]  =  0.0+1.0im
  [3, 3]  =  1.0+0.0im
  [2, 4]  =  0.0-1.0im

Qaintessent.sparse_matrix — Function

sparse_matrix(cgs::Vector{<:CircuitGate}, N::Integer=0)

returns matrix representation of a Vector{<:CircuitGate} object that can applied to a state vector of N qubits.

Examples

julia> cgs = CircuitGate[
                circuit_gate(2, Y, 1),
                circuit_gate(2, Z),
                circuit_gate(1, HadamardGate())
                ];
julia> sparse_matrix(cgs)
4×4 SparseArrays.SparseMatrixCSC{Complex{FloatQ},Int64} with 8 stored entries:
  [1, 1]  =  0.707107+0.0im
  [2, 1]  =  0.707107+0.0im
  [3, 2]  =  0.0-0.707107im
  [4, 2]  =  0.0+0.707107im
  [3, 3]  =  -0.707107+0.0im
  [4, 3]  =  -0.707107+0.0im
  [1, 4]  =  0.0-0.707107im
  [2, 4]  =  0.0+0.707107im

Compilation of abitrary unity operation

An arbitrary unitary operation can be compiled into a vector of CircuitGate objects

Qaintessent.unitary2circuit — Function

unitary2circuit(m::Matrix{Complex64,2}) where {N}

compiles quantum circuit from given unitary matrix. Implements algorithms from ff10.1016/j.cpc.2019.107001f, 10.1103/PhysRevA.69.032315, arxiv.org:1003.5760, arxiv:quant-ph/0404089

Examples

M = Stewart(ComplexF64, 4);
cgc = unitary2circuit(M)