Density Matrix representation

Qaintessent.jl supports the manipulation of mixed states as DensityMatrix(s).

Density Matrices

Qaintessent.DensityMatrix — Type

DensityMatrix

Density matrix, represented with respect to identity and Pauli basis (σ_j/2) by storing 4 real coefficients for each qubit, i.e., a real vector of length 4^N for N qubits.

Qaintessent.apply — Method

apply(cgs::Vector{<:CircuitGate}, ρ::DensityMatrix)

Apply a CircuitGate to a quantum density matrix ρ.

Qaintessent.apply — Method

apply(c::Circuit{N}, ρ::DensityMatrix) where {N}

Compute list of expectation values from measurement operators in c.meas, after applying circuit gates in c.cgc on the N-qubit density matrix ρ