Density Matrix representation

Qaintessent.jl supports the manipulation of mixed states as Density Matrices(s).

Density Matrices

Density matrices are stored as linear combination of Pauli matrices. Given a quantum state of N qubits, we require a Vector{<:Real} of length 4^N to store the mixed state.

Qaintessent.DensityMatrix — Method

DensityMatrix(v::AbstractVector{<:Real}, N::Integer)

Density matrix representation given Pauli-vector dotted with with vector of coefficients

Qaintessent.DensityMatrix — Method

DensityMatrix(v::AbstractVector{<:Real})

Density matrix representation given Pauli-vector dotted with with vector of coefficients

Several helper functions are added to crate DensityMatrix objects from quantum states in statevector form.

Qaintessent.density_from_statevector — Method

density_from_statevector(ψ::Vector)

Construct density matrix |ψ⟩⟨ψ| (Pauli representation) corresponding to quantum state ψ.

julia> ψ = 1/sqrt(2)[1, -1]; # create ket - state.
julia> ρ = density_from_statevector(ψ)
DensityMatrix([0.9999999999999998, -0.9999999999999998, 0.0, 0.0], 1)

Qaintessent.density_from_matrix — Method

density_from_matrix(ρmat::AbstractMatrix)

Construct density matrix in Pauli representation from matrix ρmat.

julia> ρmat = 0.15 .*[ 1 -1; -1  1] + 0.35 .* [1 1; 1 1]; # create mixed state of + and -.
julia> ρ = density_from_matrix(ρmat)
DensityMatrix([1.0, 0.39999999999999997, 0.0, 0.0], 1)

Once created, they can be treated similar to a statevector, see Circuit Construction and Usage.

Qaintessent.apply — Method

apply(ρ::DensityMatrix, cgs::Vector{<:CircuitGate})

Apply a CircuitGate to a quantum density matrix ρ.

Qaintessent.apply — Method

apply(ρ::DensityMatrix, c::Circuit{N}) where {N}

Compute list of expectation values from measurement operators in c.meas, after applying circuit gates in c.cgc on the N-qubit density matrix ρ