Gates

Any quantum operation can be represented as a unitary matrix, which are referred to as gates. In Qaintessent.jl, basic gates are implemented using the abstract base struct AbstractGate. The available basic quantum gates are seen below. Note that wire index information is still needed to when applying such gates. As such, these gates are permuted when creating the basic building blocks of Qaintessent.jl Quantum Circuit(s): CircuitGate(s).

Qaintessent.AbstractGate — Type

AbstractGate{N}

Abtract unitary quantum gate. N is the number of "wires" the gate acts on.

Basic Single Qubit Gates

Qaintessent.XGate — Type

Pauli X Matrix

$X = \begin{pmatrix} 0 & 1 \\ 1 & 0 \end{pmatrix}$

Qaintessent.YGate — Type

Pauli Y Matrix

$Y = \begin{pmatrix} 0 & -i \\ i & 0 \end{pmatrix}$

Qaintessent.ZGate — Type

Pauli Z Matrix

$Z = \begin{pmatrix} 1 & 0 \\ 0 & -1 \end{pmatrix}$

Qaintessent.HadamardGate — Type

Hadamard Matrix

$H = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 1 \\ 1 & 1 \end{pmatrix}$

Qaintessent.SGate — Type

S Matrix

$S = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 0 \\ 0 & i \end{pmatrix}$

Qaintessent.SdagGate — Type

S† Matrix

$S^{†} = \begin{pmatrix} 1 & 0 \\ 0 & -i \end{pmatrix}$

Qaintessent.TGate — Type

T Matrix

$T = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 0 \\ 0 & e^{\frac{iπ}{4}} \end{pmatrix}$

Qaintessent.TdagGate — Type

T† Matrix

$T^{†} = \begin{pmatrix} 1 & 0 \\ 0 & e^{-\frac{iπ}{4}} \end{pmatrix}$

Parametrized Single Qubit Gates

Qaintessent.RxGate — Type

Rotation X Matrix

$R_{x}(\theta) = \begin{pmatrix} \cos(\frac{\theta}{2}) & -i\sin(\frac{\theta}{2}) \\ -i\sin(\frac{\theta}{2}) & \cos(\frac{\theta}{2}) \end{pmatrix}$

Qaintessent.RyGate — Type

Rotation Y Matrix

$R_{y}(\theta) = \begin{pmatrix} \cos(\frac{\theta}{2}) & -\sin(\frac{\theta}{2}) \\ \sin(\frac{\theta}{2}) & \cos(\frac{\theta}{2}) \end{pmatrix}$

Qaintessent.RzGate — Type

Rotation Z Matrix

$R_{z}(\theta) = \begin{pmatrix} e^{\frac{-i\theta}{2}} & 0 \\ 0 & e^{\frac{i\theta}{2}} \end{pmatrix}$

Qaintessent.RotationGate — Type

General Rotation Matrix Rotation by angle θ around unit vector n⃗.

$R_{\vec{n}}(\theta) = \cos(\frac{\theta}{2})I - i\sin(\frac{\theta}{2})\vec{n}\sigma, \\ \sigma = [X, Y, Z]$

Qaintessent.PhaseShiftGate — Type

Phase Shift Gate

$P(\phi) = \begin{pmatrix} 1 & 0 \\ 0 & e^{i\phi} \end{pmatrix}$

Two Qubit Gates

Qaintessent.SwapGate — Type

Swap Gate

$SWAP = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & 1 \end{pmatrix}$

Quantum Gate Matrix Representation

The Qaintessent.matrix function can be used to convert AbstractGate objects to their matrix representation.

using Qaintessent

x = XGate()
Qaintessent.matrix(x)

2×2 Array{Float64,2}:
 0.0  1.0
 1.0  0.0