Gates

Any quantum operation can be represented as a unitary matrix, which are referred to as gates. In Qaintessent.jl, basic gates are implemented using the abstract struct AbstractGate. The available basic quantum gates are seen below. Note that at this level of abstraction, wire index information is ignored and matrixes are permuted such that the fastest running qubit is also the fastest running index.

These gates are permuted when creating the basic building blocks of Qaintessent.jl Quantum Circuit(s): CircuitGate(s).

Qaintessent.AbstractGate — Type

AbstractGate

Abtract unitary quantum gate. N is the number of "wires" the gate acts on.

AbstractGate helper functions

Qaintessent.matrix — Method

matrix(::AbstractGate)

returns dense matrix representation of AbstractGate objects

Examples

julia> matrix(X)
2×2 Array{Complex{Float64},2}:
 0.0+0.0im  1.0+0.0im
 1.0+0.0im  0.0+0.0im

Qaintessent.sparse_matrix — Method

sparse_matrix(::AbstractGate)

returns sparse matrix representation of AbstractGate objects

Examples

julia> sparse_matrix(X)
2×2 SparseArrays.SparseMatrixCSC{Complex{Float64},Int64} with 2 stored entries:
  [2, 1]  =  1.0+0.0im
  [1, 2]  =  1.0+0.0im

Base.adjoint — Method

Base.adjoint(::AbstractGate)

returns adjoint of AbstractGate objects

Examples

julia> X ≈ X
true

julia> Y ≈ Z
false

julia> RxGate(0.3) ≈ RxGate(0.3)
true

julia> RxGate(0.3) ≈ RxGate(0.2)
false

Qaintessent.num_wires — Method

num_wires(::AbstractGate)

returns number of wires affected by AbstractGate object.

Examples

julia> num_wires(X)
1

julia> num_wires(SwapGate())
2

Base.isapprox — Method

Base.isapprox(g1::G, g2::G) where {G <: AbstractGate}

returns true if input AbstractGates are of same type with same paremeters

Examples

julia> X ≈ X
true

julia> X ≈ Y
false

julia> RxGate(0.1) ≈ RxGate(0.2)
false

Base.kron — Method

kron(m::Union{UniformScaling{Bool},G}...) where {G<:AbstractGate}

returns sparse matrix representation of kronecker product of AbstractGate objects

Basic Single Qubit Gates

Qaintessent.XGate — Type

Pauli X gate

$X = \begin{pmatrix} 0 & 1 \\ 1 & 0 \end{pmatrix}$

Qaintessent.YGate — Type

Pauli Y gate

$Y = \begin{pmatrix} 0 & -i \\ i & 0 \end{pmatrix}$

Qaintessent.ZGate — Type

Pauli Z gate

$Z = \begin{pmatrix} 1 & 0 \\ 0 & -1 \end{pmatrix}$

Qaintessent.HadamardGate — Type

Hadamard gate

$H = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 1 \\ 1 & 1 \end{pmatrix}$

Qaintessent.SGate — Type

S gate

$S = \begin{pmatrix} 1 & 0 \\ 0 & i \end{pmatrix}$

Qaintessent.SdagGate — Type

S^{†} gate

$S^{†} = \begin{pmatrix} 1 & 0 \\ 0 & -i \end{pmatrix}$

Qaintessent.TGate — Type

T gate

$T = \begin{pmatrix} 1 & 0 \\ 0 & e^{\frac{iπ}{4}} \end{pmatrix}$

Qaintessent.TdagGate — Type

T^{†} gate

$T^{†} = \begin{pmatrix} 1 & 0 \\ 0 & e^{-\frac{iπ}{4}} \end{pmatrix}$

Basic Two Qubit Gates

Qaintessent.SwapGate — Type

Swap gate

$SWAP = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & 1 \end{pmatrix}$

Parametrized Gates

Qaintessent.RxGate — Type

Rotation-X gate

$R_{x}(\theta) = \begin{pmatrix} \cos(\frac{\theta}{2}) & -i\sin(\frac{\theta}{2}) \\ -i\sin(\frac{\theta}{2}) & \cos(\frac{\theta}{2}) \end{pmatrix}$

Qaintessent.RyGate — Type

Rotation-Y gate

$R_{y}(\theta) = \begin{pmatrix} \cos(\frac{\theta}{2}) & -\sin(\frac{\theta}{2}) \\ \sin(\frac{\theta}{2}) & \cos(\frac{\theta}{2}) \end{pmatrix}$

Qaintessent.RzGate — Type

Rotation-Z gate

$R_{z}(\theta) = \begin{pmatrix} e^{\frac{-i\theta}{2}} & 0 \\ 0 & e^{\frac{i\theta}{2}} \end{pmatrix}$

Qaintessent.RotationGate — Type

General rotation operator gate: rotation by angle `θ` around unit vector `n`.

$R_{\vec v}(\theta) = \cos(\frac{\theta}{2})I - i\sin(\frac{\theta}{2})\vec v \sigma, \\ \sigma = [X, Y, Z]$

Qaintessent.PhaseShiftGate — Type

Phase shift gate

$P(\phi) = \begin{pmatrix} 1 & 0 \\ 0 & e^{i\phi} \end{pmatrix}$

Qaintessent.EntanglementXXGate — Type

Entanglement-XX gate

$G_{x}(\theta) = e^{-i \theta X \otimes X / 2}$

Reference:

B. Kraus and J. I. Cirac

Optimal creation of entanglement using a two-qubit gate

Phys. Rev. A 63, 062309 (2001)

Qaintessent.EntanglementYYGate — Type

Entanglement-YY gate

$G_{y}(\theta) = e^{-i \theta Y \otimes Y / 2}$

Reference:

B. Kraus and J. I. Cirac

Optimal creation of entanglement using a two-qubit gate

Phys. Rev. A 63, 062309 (2001)

Qaintessent.EntanglementZZGate — Type

Entanglement-ZZ gate

$G_{z}(\theta) = e^{-i \theta Z \otimes Z / 2}$

Reference:

B. Kraus and J. I. Cirac

Optimal creation of entanglement using a two-qubit gate

Phys. Rev. A 63, 062309 (2001)

Controlled Gates

Arbitrary controlled gates using the basic gates defined above are available.

Qaintessent.ControlledGate — Type

ControlledGate{G}(U::G, M::Int) where {G<:AbstractGate}

returns ControlledGate object. Controls AbstractGate of type G with M control wires

ControlledGate helper functions

Qaintessent.control_wires — Function

control_wires(g::ControlledGate)

return control wires for a ControlledGate

Examples

# controlled gate with 2 control wires
julia> ccnot = ControlledGate(X, 2);
julia> target_wires(ccnot)
2

Qaintessent.target_wires — Function

target_wires(g::ControlledGate)

return target wires for a ControlledGate

Examples

# controlled gate with 1 target wire
julia> cnot = ControlledGate(X, 1);
julia> target_wires(cnot)
1