Models

Qaintessent.qft_circuit — Method

qft_circuit(N)

Construct the quantum Fourier transform circuit for N qubits.

Missing docstring.

Missing docstring for toffoli_circuit(cntrl::Tuple{<:Integer, <:Integer} , trg::Integer, N::Integer). Check Documenter's build log for details.

Qaintessent.vbe_adder_circuit — Method

vbe_adder_circuit(N)

Construct an in-place adder for 2 integers represented by N qubits. Based on ripple-carry adder circuit in Vedral et. al (Phys. Rev. A 54, 147 (1996), arXiv:quant-ph/9511018) Returns a CircuitGateChain{3N+1} as there are N+1 ancillary wires

If the two added integers are represented as:

$A = a_{0}\times 2^{0} + a_{1} \times 2^{1} + a_{2} \times 2^{2} + .. + a_{N} \times 2^{N} \\ B = b_{0}\times 2^{0} + b_{1} \times 2^{1} + b_{2} \times 2^{2} + .. + b_{N} \times 2^{N}$

The input index should be $a_{0}a_{1}a_{2}..a_{N}b_{0}b_{1}b_{2}..b_{N} + 1$ with $a_{0}$ as the fastest running index

The output index will be in the form $a_{0}a_{1}a_{2}..a_{N}c_{0}c_{1}c_{2}..c_{N} + 1$ where:

$C = (A+B) \% (2^{N} + 1) = c_{0}\times 2^{0} + c_{1} \times 2^{1} + c_{2} \times 2^{2} + ... + c_{N} \times 2^{N}$

Example

using Qaintessent 

N = 3
M = N*3 + 1

adder = vbe_adder_circuit(N)
ψ = fill(0.0+0.0*im, 2^M)
a = 1
b = 3

index = b << N + a
ψ[index+1] = 1.0

ψ = apply(adder, ψ)
((findall(x->x==1, ψ)[1]-1)%(2^2N)) >> N

# output

4

Qaintessent.qcla_out_adder_circuit — Method

qcla_out_adder_circuit(N)

Construct an out-of-place adder for 2 integers represented by N qubits. returns a CircuitGateChain{3N+1} object. Based on quantum carry-lookahead adder circuit by Draper et. al (Quant. Inf. Comp. 6, 351-369 (2006), arXiv:quant-ph/0406142) Returns a CircuitGateChain{3N+1} as there are N+1 ancillary wires If the two added integers are represented as: $A = a_{0}\times 2^{0} + a_{1} \times 2^{1} + a_{2} \times 2^{2} + .. + a_{N} \times 2^{N} \\ B = b_{0}\times 2^{0} + b_{1} \times 2^{1} + b_{2} \times 2^{2} + .. + b_{N} \times 2^{N}$ The input index should be $a_{0}a_{1}a_{2}..a_{N}b_{0}b_{1}b_{2}..b_{N} + 1$ as Julia starts indexing at 1 and $a_{0}$ as the fastest running index The output index will be in the form $a_{0}a_{1}a_{2}..a_{N}b_{0}b_{1}b_{2}..b_{N}c_{0}c_{1}c_{2}..c_{N+1} + 1$ where: $C = A+B = c_{0}\times 2^{0} + c_{1} \times 2^{1} + c_{2} \times 2^{2} + ... + c_{N+1} \times 2^{N+1}$

Example

using Qaintessent 

N = 3
M = 3N + 1

adder = qcla_out_adder_circuit(N)
ψ = fill(0.0+0.0*im, 2^M)
a = 2
b = 3

index = b << N + a
ψ[index+1] = 1.0

ψ = apply(adder, ψ)
(findall(x->x==1, ψ)[1] - 1) >> 2N

# output

5

Qaintessent.qcla_inplace_adder_circuit — Method

qcla_inplace_adder_circuit(N)

Construct an in-place adder for 2 integers represented by N qubits. Based on quantum carry-lookahead adder circuit by Draper et. al (Quant. Inf. Comp. 6, 351-369 (2006), arXiv:quant-ph/0406142) Returns a CircuitGateChain{3N+1} as there are N+1 ancillary wires

If the two added integers are represented as:

$A = a_{0}\times 2^{0} + a_{1} \times 2^{1} + a_{2} \times 2^{2} + .. + a_{N} \times 2^{N} \\ B = b_{0}\times 2^{0} + b_{1} \times 2^{1} + b_{2} \times 2^{2} + .. + b_{N} \times 2^{N}$

The input index should be $a_{0}a_{1}a_{2}..a_{N}b_{0}b_{1}b_{2}..b_{N} + 1$ as Julia starts indexing at 1 and $a_{0}$ as the fastest running index

The output index will be in the form $a_{0}a_{1}a_{2}..a_{N}c_{0}c_{1}c_{2}..c_{N+1} + 1$ where:

$C = A+B = c_{0}\times 2^{0} + c_{1} \times 2^{1} + c_{2} \times 2^{2} + ... + c_{N+1} \times 2^{N+1}$

Example

using Qaintessent 

N = 3
M = 3N

adder = qcla_inplace_adder_circuit(N)
ψ = fill(0.0+0.0*im, 2^M)
a = 1
b = 2

index = b << N + a
ψ[index+1] = 1.0

ψ = apply(adder, ψ)
(findall(x->x==1, ψ)[1] - 1) >> N

# output

3